+++ src/main/java/com/example/braveCoward/global/exectime/ExecutionTimeLogger.java
@@ -1,0 +1,11 @@
+package com.example.braveCoward.global.exectime;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ExecutionTimeLogger {
+}
