+++ src/main/java/com/example/braveCoward/global/extractuserid/ExtractAuthenticationInterceptor.java
@@ -1,4 +1,4 @@
-package com.example.braveCoward.global;
+package com.example.braveCoward.global.extractuserid;
 
 import java.util.Optional;
