+++ src/main/java/com/example/braveCoward/global/extractuserid/UserIdContext.java
@@ -1,23 +1,18 @@
-package com.example.braveCoward.global;
+package com.example.braveCoward.global.extractuserid;
+
+import java.util.Objects;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.annotation.RequestScope;
 
 @Component
 @RequestScope
-public class AuthContext {
+public class UserIdContext {
 
     private Integer userId;
 
     public Integer getUserId() {
-        if (userId == null) {
-            throw new IllegalArgumentException("Access token is required or invalid.");
-        }
         return userId;
-    }
-
-    public boolean isAnonymous() {
-        return userId == null;
     }
 
     public void setUserId(Integer userId) {
